+++ pkgs/messages/example/bin/example.dart
@@ -6,7 +6,7 @@
 
 import 'dart:io';
 
-import 'package:example/AboutPage_messages.g.dart';
+import 'package:messages_example/AboutPage_messages.g.dart';
 
 Future<void> main(List<String> arguments) async {
   final messages = AboutPageMessages(
@@ -18,19 +18,13 @@ Future<void> main(List<String> arguments) async {
   print('AboutMessage en:');
   print('\t${messages.aboutMessage('typesafe.en')}');
 
-  ///To enable this, add `generateFindById: true` to the pubspec section
-  // print('\t${messages.getById('aboutMessage', ['get-by-id.en'])}');
-
-  ///To enable this, add `generateFindBy: enumerate` to the pubspec section
-  // print('\t${messages.getByEnum(index, ['get-by-index.en'])}');
-
   await messages.loadLocale('fr');
   print('AboutMessage fr:');
   print('\t${messages.aboutMessage('typesafe.fr')}');
+  print(getGreeting('Dart'));
+}
 
-  ///To enable this, add `generateFindById: true` to the pubspec section
-  // print('\t${messages.getById('aboutMessage', ['get-by-id.fr'])}');
-
-  ///To enable this, add `generateFindBy: enumerate` to the pubspec section
-  // print('\t${messages.getByEnum(index, ['get-by-index.fr'])}');
+String getGreeting(String name) {
+  // Try clicking on the string literal below to test "Extract to ARB"!
+  return 'Welcome back, $name!';
 }