+++ pkgs/messages_builder/lib/code_generation/import_generation.dart
@@ -15,14 +15,14 @@ class ImportGeneration {
   List<Directive> generate() {
     final serializationImports = switch (options.deserialization) {
       DeserializationType.web => [
-          Directive.import('package:messages/messages_json.dart')
-        ],
+        Directive.import('package:messages/messages_json.dart'),
+      ],
     };
     final pluralImports = switch (options.pluralSelector) {
       PluralSelectorType.intl => [Directive.import('package:intl/intl.dart')],
       PluralSelectorType.intl4x => [
-          Directive.import('package:intl4x/intl4x.dart')
-        ],
+        Directive.import('package:intl4x/plural_rules.dart'),
+      ],
       PluralSelectorType.custom => <Directive>[],
     };
 
@@ -30,9 +30,15 @@ class ImportGeneration {
       return Directive.importDeferredAs('$emptyFilePath.g.dart', emptyFilePath);
     });
 
+    final targetImports = [
+      if (options.assetLoadingStyle == TargetEnvironment.flutter)
+        Directive.import('package:flutter/services.dart', show: ['rootBundle']),
+    ];
+
     return [
       ...serializationImports,
       ...pluralImports,
+      ...targetImports,
       ...deferredImports,
     ];
   }