+++ pkgs/messages_builder/test/web_deserializer_native_test.dart
@@ -42,77 +42,69 @@ void main() {
     final buffer = JsonSerializer()
         .serialize('', '', messageList.map((e) => e.message).toList())
         .data;
-    final messages =
-        JsonDeserializer(buffer).deserialize(intlPluralSelector).messages;
+    final messages = JsonDeserializer(buffer).deserialize().messages;
     expect((messages[0] as StringMessage).value, message.value);
   });
 
   test('generateMessageFile from simple arb JSON', () {
-    final arb = <String, dynamic>{
+    final arb = <String, Object?>{
       '@@locale': 'en',
-      'helloWorld': 'Hello World'
+      'helloWorld': 'Hello World',
     };
     final parsed = ArbParser().parseMessageFile(arb);
     final buffer = JsonSerializer()
         .serialize('', '', parsed.messages.map((e) => e.message).toList())
         .data;
-    final messages =
-        JsonDeserializer(buffer).deserialize(intlPluralSelector).messages;
+    final messages = JsonDeserializer(buffer).deserialize().messages;
     expect((messages[0] as StringMessage).value, 'Hello World');
   });
   test('generateMessageFile from simple arb JSON with placeholder', () {
-    final arb = <String, dynamic>{
+    final arb = <String, Object?>{
       '@@locale': 'en',
-      'helloWorld': 'Hello {name}'
+      'helloWorld': 'Hello {name}',
     };
     final parsed = ArbParser().parseMessageFile(arb);
     final buffer = JsonSerializer()
         .serialize('', '', parsed.messages.map((e) => e.message).toList())
         .data;
-    final messages =
-        JsonDeserializer(buffer).deserialize(intlPluralSelector).messages;
+    final messages = JsonDeserializer(buffer).deserialize().messages;
     expect((messages[0] as StringMessage).value, 'Hello ');
-    expect(
-      (messages[0] as StringMessage).argPositions,
-      [(argIndex: 0, stringIndex: 6)],
-    );
+    expect((messages[0] as StringMessage).argPositions, [
+      (argIndex: 0, stringIndex: 6),
+    ]);
   });
   test('generateMessageFile from simple arb JSON with only placeholders', () {
-    final arb = <String, dynamic>{
+    final arb = <String, Object?>{
       '@@locale': 'en',
-      'helloWorld': '{greeting}{space}{name}'
+      'helloWorld': '{greeting}{space}{name}',
     };
     final parsed = ArbParser().parseMessageFile(arb);
     final buffer = JsonSerializer()
         .serialize('', '', parsed.messages.map((e) => e.message).toList())
         .data;
-    final messages =
-        JsonDeserializer(buffer).deserialize(intlPluralSelector).messages;
+    final messages = JsonDeserializer(buffer).deserialize().messages;
     expect((messages[0] as StringMessage).value, '');
-    expect(
-      (messages[0] as StringMessage).argPositions,
-      [
-        (argIndex: 0, stringIndex: 0),
-        (argIndex: 1, stringIndex: 0),
-        (argIndex: 2, stringIndex: 0),
-      ],
-    );
+    expect((messages[0] as StringMessage).argPositions, [
+      (argIndex: 0, stringIndex: 0),
+      (argIndex: 1, stringIndex: 0),
+      (argIndex: 2, stringIndex: 0),
+    ]);
   });
 
   test('generateMessageFile from complex arb JSON', () {
-    final arb = jsonDecode(arbFile) as Map<String, dynamic>;
+    final arb = jsonDecode(arbFile) as Map<String, Object?>;
     final parsed = ArbParser().parseMessageFile(arb);
     final buffer = JsonSerializer()
         .serialize('', '', parsed.messages.map((e) => e.message).toList())
         .data;
-    final messages =
-        JsonDeserializer(buffer).deserialize(intlPluralSelector).messages;
+    final messages = JsonDeserializer(buffer).deserialize().messages;
     expect(
-        messages[2].generateString(
-          ['female', 'b'],
-          pluralSelector: intlPluralSelector,
-        ),
-        'test One new message');
+      messages[2].generateString([
+        'female',
+        'b',
+      ], pluralSelector: intlPluralSelector),
+      'test One new message',
+    );
   });
 
   test('Key with spaces is not ok', () {