+++ pkgs/messages_serializer/lib/src/serializer_json.dart
@@ -9,9 +9,9 @@ import 'package:messages/messages_json.dart';
 import 'serializer.dart';
 
 class JsonSerializer extends Serializer<String> {
-  final List result = [];
+  final List<Object?> result = [];
 
-  JsonSerializer([super.writeIds = false]);
+  JsonSerializer();
 
   @override
   Serialization<String> serialize(
@@ -26,39 +26,37 @@ class JsonSerializer extends Serializer<String> {
       serializationVersion: serializationVersion,
       locale: locale,
       hash: hash,
-      hasIds: writeIds,
     );
 
     result.addAll(preamble.toJson());
 
-    final messageMapping = <String, String>{};
-    var messageCounter = 0;
+    final keptIndices = <int>[];
     for (var i = 0; i < messages.length; i++) {
       if (keepOnly?.contains(i) ?? true) {
         encodeMessage(messages[i], isVisible: true);
-        messageMapping[i.toRadixString(serializationRadix)] =
-            messageCounter.toRadixString(serializationRadix);
-        messageCounter++;
+        if (keepOnly != null) {
+          keptIndices.add(i);
+        }
       }
     }
 
-    /// Insert `null` instead of the full messageMapping to save space.
-    result.insert(Preamble.length, keepOnly != null ? messageMapping : null);
+    /// Insert `null` instead of keptIndices to save space when all messages are
+    /// kept.
+    result.insert(Preamble.length, keepOnly != null ? keptIndices : null);
 
     return Serialization(jsonEncode(result));
   }
 
   Object encodeMessage(Message message, {bool isVisible = false}) {
-    // print('Encode message $message');
     Object messageIndex;
     if (message is StringMessage) {
-      messageIndex = encodeString(message, isVisible);
+      messageIndex = encodeString(message);
     } else if (message is SelectMessage) {
-      messageIndex = encodeSelect(message, isVisible);
+      messageIndex = encodeSelect(message);
     } else if (message is PluralMessage) {
-      messageIndex = encodePlural(message, isVisible);
+      messageIndex = encodePlural(message);
     } else if (message is CombinedMessage) {
-      messageIndex = encodeCombined(message, isVisible);
+      messageIndex = encodeCombined(message);
     } else {
       throw ArgumentError('Unknown message type');
     }
@@ -70,28 +68,23 @@ class JsonSerializer extends Serializer<String> {
 
   /// Encodes a string message as follows:
   ///
-  /// If the id does not have to be written, and there are no placeholders:
+  /// If there are no placeholders:
   /// * the String value
   /// else:
-  /// * int | the StringMessage type
-  /// * if we write IDs: String | the message id
   /// * String | the String value
-  /// * if there are placeholders: List\<List\> | the position pairs:
+  /// * List\<List\> | the position pairs:
   ///   * List\<int\> | a pair of position in the string - number of the placeholder
-  Object encodeString(StringMessage message, bool isVisible) {
+  Object encodeString(StringMessage message) {
     final containsArgs = message.argPositions.isNotEmpty;
-    if ((message.id == null || isVisible == false) && !containsArgs) {
+    if (!containsArgs) {
       return message.value;
     }
     final m = <Object>[];
-    addId(message, m, isVisible);
     m.add(message.value);
-    if (containsArgs) {
-      final positions = message.argPositions
-        ..sort((a, b) => a.stringIndex.compareTo(b.stringIndex));
-      for (var i = 0; i < positions.length; i++) {
-        m.add(<int>[positions[i].stringIndex, positions[i].argIndex]);
-      }
+    final positions = List.of(message.argPositions)
+      ..sort((a, b) => a.stringIndex.compareTo(b.stringIndex));
+    for (var i = 0; i < positions.length; i++) {
+      m.add(<int>[positions[i].stringIndex, positions[i].argIndex]);
     }
     return m;
   }
@@ -99,14 +92,12 @@ class JsonSerializer extends Serializer<String> {
   /// Encodes a select message as follows:
   ///
   /// * int | the SelectMessage type
-  /// * if we write IDs: String | the message id
   /// * int | the argument index on which the select switches
   /// * Map\<String, int\> | the cases:
   ///   * MapEntry\<String, int\> | a case mapped to the message it represents
-  List encodeSelect(SelectMessage message, bool isVisible) {
+  List<Object> encodeSelect(SelectMessage message) {
     final m = <Object>[];
     m.add(SelectMessage.type);
-    addId(message, m, isVisible);
     m.add(message.argIndex);
     m.add(encodeMessage(message.other));
     final caseIndices = <String, Object>{};
@@ -120,16 +111,14 @@ class JsonSerializer extends Serializer<String> {
   /// Encodes a plural message as follows:
   ///
   /// * int | the PluralMessage type
-  /// * if we write IDs: String | the message id
   /// * int | the argument index on which the plural switches
   /// * int | the index of the other case message, which must be present
   /// * List\<int\> | the cases, which are added in pairs of two:
   ///   * int | the case index as encoded by the constants in `Plural`
   ///   * int | the message index of the case
-  List encodePlural(PluralMessage message, bool isVisible) {
+  List<Object> encodePlural(PluralMessage message) {
     final m = <Object>[];
     m.add(PluralMessage.type);
-    addId(message, m, isVisible);
     m.add(message.argIndex);
     m.add(encodeMessage(message.other));
     final caseIndices = <Object>[];
@@ -157,25 +146,18 @@ class JsonSerializer extends Serializer<String> {
   /// Encodes a combined message as follows:
   ///
   /// * int | the CombinedMessage type
-  /// * if we write IDs: String | the message id
   /// * List\<int\> | the submessage IDs
   ///   * int | the index of the submessage
-  List encodeCombined(CombinedMessage message, bool isVisible) {
+  List<Object> encodeCombined(CombinedMessage message) {
     final m = <Object>[];
     m.add(CombinedMessage.type);
-    addId(message, m, isVisible);
     for (var submessage in message.messages) {
       m.add(encodeMessage(submessage));
     }
     return m;
   }
 
-  /// Add a non-null ID iff `writeIds` is enabled
-  void addId(Message message, List<dynamic> m, bool isVisible) {
-    if (writeIds && message.id != null && isVisible) m.add(message.id!);
-  }
-
-  int addMessage(dynamic m) {
+  int addMessage(Object m) {
     result.add(m);
     return result.length - 1;
   }